INTRODUCTION-TO-RIP
Routing Information Protocol (RIP) is a dynamic routing protocol that uses hop count as a routing metric to find the best path between the source and the destination network. It is a distance-vector routing protocol that has an AD value of 120 and works on the Network layer of the OSI model. RIP uses port number 520. 

Best source: https://www.geeksforgeeks.org/routing-information-protocol-rip/