https://www.cepro.com/news/networking_geeks_why_router_on_a_stick_is_good
What many people fail to realize is that there are some pretty huge benefits to using a router, or more precisely a firewall, to be the gateway between networks.

To understand the benefits of putting a firewall in-between segments you must understand something that many people overlook: Most internal traffic never touches the router interface unless it’s going out to the internet.

In other words, any firewall access rules, traffic-shaping, anti-malware scanning, etc., cannot be used on the internal network UNLESS it hits a router interface.

This is the reason IP cameras can communicate with an NVR without even needing a gateway address: You can physically remove the router and they’ll still be able to communicate with each other. Sure, they won’t be accessible from the Internet, but they can still communicate with one another.

If you had a firewall in place there is nothing you could do with that firewall to keep them from communicating. There is essentially zero security that a firewall would provide in this scenario. The traffic must hit the router interface in order for the firewall to do its thing.

Now, if you want to be able to use the various features of a full-fledged firewall, then the router-on-a-stick method starts to look a whole lot better. Some of these features include:

shaping traffic between VLANs
segmenting traffic with more granularity than simple access rules that most switches can provide
scanning traffic between networks for malware or spyware, or any other
packet monitoring